+++ types/Elements.types.ts
@@ -1,3 +1,8 @@
+export interface BaseButtonProps {
+  onClick: () => void;
+}
+
+
 export interface ButtonProps {
 	className?: string;
 	label?: string;